moved test files in a separate folder.
